+++ src/Query/InsertQueryBuilder.php
@@ -74,4 +74,14 @@ class InsertQueryBuilder
     {
         return $this->queryExecutor->insert($this->queryBuilder);
     }
+
+    public function getSql(): string
+    {
+        return $this->queryExecutor->getSql($this->queryBuilder);
+    }
+
+    public function getParams(): array
+    {
+        return $this->queryExecutor->getParams($this->queryBuilder);
+    }
 }