+++ src/Query/SelectQueryBuilder.php
@@ -223,6 +223,16 @@ class SelectQueryBuilder
         return $this->queryExecutor->fetchColumn($this->queryBuilder);
     }
 
+    public function getSql(): string
+    {
+        return $this->queryExecutor->getSql($this->queryBuilder);
+    }
+
+    public function getParams(): array
+    {
+        return $this->queryExecutor->getParams($this->queryBuilder);
+    }
+
     private function getModelRelationships(ModelInterface $model): array
     {
         $result = [];