+++ src/Query/UpdateQueryBuilder.php
@@ -80,4 +80,14 @@ class UpdateQueryBuilder
 
         return $this->queryExecutor->update($this->queryBuilder);
     }
+
+    public function getSql(): string
+    {
+        return $this->queryExecutor->getSql($this->queryBuilder);
+    }
+
+    public function getParams(): array
+    {
+        return $this->queryExecutor->getParams($this->queryBuilder);
+    }
 }